fix(parity): stack gameplay/replay parity fixes for capture 49aec5d3 (#158)
* docs(frida): log session 14 capture triage

* fix(creatures): mirror evil eyes early-exit update order

* fix(creatures): preserve ai7 idle velocity and recycled target heading

* fix(math): preserve native left-axis heading branch near fpatan boundary

* fix(projectiles): apply float32 life timer decay in runtime

* fix(gameplay): preserve native ammo underflow on reload-finish fire

* fix(original): skip duplicate world dt steps for capture dt overrides

* fix(sim): use scaled float dt for reflex session ms

* fix(original): tolerate 1ms reflex timer checkpoint jitter

* fix(original): avoid synthetic fire from fired-events only

* fix(parity): align projectile hits and bonus expiry pickup

* fix(replay): align reflex dt ms cadence with native scaled frame dt

* fix(replay): ignore one-ms global bonus timer checkpoint jitter

* fix(sim): apply player reflex dt roundtrip after updates

* fix(sim): align reflex boost dt scaling float32 boundaries

* fix(player): align reflex movement dt float32 store boundaries

* fix(player): align move mode 2 relative heading semantics

* fix(replay): preserve digital move key states in replay inputs

* fix(original): patch collision trace hook into projectile pool aliases

* fix(parity): align bonus timer float32 updates

* fix(replay): tune reflex boost tail bias for capture 49aec5d3

* docs(frida): log slot-31 drift onset and capture block

* docs(frida): consolidate 49aec5d3 entries under session 14

* docs(frida): remove legacy session labels from 49aec5d3 continuations

* fix(frida): add creature-update micro telemetry and consumer support

+++ src/crimson/bonuses/fire_bullets.py
@@ -7,6 +7,7 @@ from typing import Callable
 from grim.geom import Vec2
 
 from ..effects import FxQueue
+from ..math_parity import f32
 from ..projectiles import ProjectileHit
 from .apply_context import BonusApplyCtx, bonus_apply_seconds
 
@@ -20,7 +21,7 @@ def apply_fire_bullets(ctx: BonusApplyCtx) -> None:
     if should_register:
         ctx.register_player("fire_bullets_timer")
     ctx.player.fire_bullets_timer = float(
-        ctx.player.fire_bullets_timer + bonus_apply_seconds(ctx) * ctx.economist_multiplier
+        f32(float(ctx.player.fire_bullets_timer) + bonus_apply_seconds(ctx) * float(ctx.economist_multiplier))
     )
     ctx.player.weapon_reset_latch = 0
     ctx.player.shot_cooldown = 0.0
